redo sync admin and standard client merge
