--- .github/workflows/extensive_vector_search_tests.yml
@@ -1,81 +1,0 @@
-name: Run long running vector search tests
-
-on:
-  push:
-    branches:
-      - main
-
-jobs:
-  test-exhaustive-vector-search:
-    runs-on: ubuntu-24.04
-    continue-on-error: false
-
-
-    strategy:
-      matrix:
-        python-version: ["3.12"]
-
-        
-    steps:
-    - name: Checkout code
-      uses: actions/checkout@v4
-
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python-version }}
-
-
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python setup.py
-        pip install -r requirements.txt
-      working-directory: tests
-
-
-    - name: Retrieve the secret and decode it to a file
-      env:
-        FEATURE_FILE: ${{ secrets.FEATURE_FILE }}
-      run: |
-        echo $FEATURE_FILE | base64 --decode > features.conf 
-      working-directory: tests
-
-    - name: Docker Login
-      uses: docker/login-action@v2
-      with:
-        registry: aerospike.jfrog.io
-        username: ${{ secrets.JFROG_USERNAME }}
-        password: ${{ secrets.JFROG_PASSWORD }}
-
-
-    - name: Set up RANDFILE environment variable
-      run: echo "RANDFILE=$HOME/.rnd" >> $GITHUB_ENV
-
-    - name: Create .rnd file if it doesn't exist
-      run: touch $HOME/.rnd
-
-    - name: create config
-      run: |
-        assets/call_gen.sh
-        cat /etc/hosts
-      working-directory: tests
-
-    - name: Run unit tests
-      run: |
-
-        docker run -d --name aerospike-vector-search --network=host -p 5000:5000  -v $(pwd):/etc/aerospike-vector-search aerospike/aerospike-vector-search:1.0.0
-        docker run -d --name aerospike -p 3000:3000 -v .:/etc/aerospike aerospike/aerospike-server-enterprise:latest 
-
-        sleep 5
-
-        python -m pytest standard -s --host 0.0.0.0 --port 5000 --extensive_vector_search -vs
-
-        mv .coverage coverage_data
-      working-directory: tests
-
-    - name: Upload test coverage
-      uses: actions/upload-artifact@v4
-      with:
-        name: coverage_exhaustive_vector_search
-        path: tests/coverage_data
