+++ docs/aio.rst
@@ -8,5 +8,4 @@ This module contains clients with coroutine methods used for asynchronous progra
    :maxdepth: 2
    :caption: Contents:
 
-   aio_admin
    aio_client
