+++ docs/index.rst
@@ -5,14 +5,11 @@
 
 Welcome to Aerospike Vector Search Client for Python.
 
-This package splits the client functionality into two separate clients.
-
-This standard client (Client) specializes in performing database operations with vector data.
-Moreover, the standard client supports Hierarchical Navigable Small World (HNSW) vector searches,
+This client (Client) specializes in performing database operations with vector data.
+Moreover, the client supports Hierarchical Navigable Small World (HNSW) vector searches,
 allowing users to find vectors similar to a given query vector within an index.
-
-This admin client (AdminClient) is designed to conduct AVS administrative operation such
-as creating indexes, querying index information, and dropping indexes.
+administrative operations  such as creating indexes,
+querying index information, and dropping indexes are also supported.
 
 Please explore the modules below for more information on API usage and details.
