+++ pyproject.toml
@@ -22,7 +22,7 @@ classifiers = [
     "Programming Language :: Python :: Implementation :: CPython",
     "Topic :: Database"
 ]
-version = "3.1.0"
+version = "4.0.0"
 requires-python = ">3.8"
 dependencies = [
     "grpcio == 1.68.1",
