+++ src/aerospike_vector_search/__init__.py
@@ -1,5 +1,4 @@
 from .client import Client
-from .admin import Client as AdminClient
 from .types import (
     HostPort,
     Key,
