+++ tests/rbac/sync/conftest.py
@@ -1,7 +1,6 @@
 import pytest
 
 from aerospike_vector_search import Client
-from aerospike_vector_search.admin import Client as AdminClient
 from aerospike_vector_search import types
 
 
@@ -29,7 +28,7 @@ def drop_all_indexes(
         with open(private_key, "rb") as f:
             private_key = f.read()
 
-    with AdminClient(
+    with Client(
         seeds=types.HostPort(host=host, port=port),
         is_loadbalancer=is_loadbalancer,
         username=username,
@@ -47,7 +46,7 @@ def drop_all_indexes(
 
 
 @pytest.fixture(scope="module")
-def session_rbac_admin_client(
+def session_rbac_client(
     username,
     password,
     root_certificate,
@@ -70,7 +69,7 @@ def session_rbac_admin_client(
         with open(private_key, "rb") as f:
             private_key = f.read()
 
-    client = AdminClient(
+    client = Client(
         seeds=types.HostPort(host=host, port=port),
         is_loadbalancer=is_loadbalancer,
         username=username,
