+++ tests/standard/test_admin_client_index_get_status.py
@@ -13,25 +13,25 @@ import grpc
 @pytest.mark.parametrize("empty_test_case", [None])
 #@given(random_name=index_strategy())
 #@settings(max_examples=1, deadline=1000)
-def test_index_get_status(session_admin_client, empty_test_case, index):
-    result = session_admin_client.index_get_status(namespace=DEFAULT_NAMESPACE, name=index)
+def test_index_get_status(session_vector_client, empty_test_case, index):
+    result = session_vector_client.index_get_status(namespace=DEFAULT_NAMESPACE, name=index)
 
     assert result.unmerged_record_count == 0
-    drop_specified_index(session_admin_client, DEFAULT_NAMESPACE, index)
+    drop_specified_index(session_vector_client, DEFAULT_NAMESPACE, index)
 
 
 @pytest.mark.parametrize("empty_test_case", [None])
 #@given(random_name=index_strategy())
 #@settings(max_examples=1, deadline=1000)
 def test_index_get_status_timeout(
-    session_admin_client, empty_test_case, index, with_latency
+    session_vector_client, empty_test_case, index, with_latency
 ):
     if not with_latency:
         pytest.skip("Server latency too low to test timeout")
 
     for i in range(10):
         try:
-            result = session_admin_client.index_get_status(
+            result = session_vector_client.index_get_status(
                 namespace=DEFAULT_NAMESPACE, name=index, timeout=0.0001
             )
         except AVSServerError as se:
