+++ tests/standard/test_vector_client_is_indexed.py
@@ -8,14 +8,13 @@ import grpc
 
 
 def test_vector_is_indexed(
-    session_admin_client,
     session_vector_client,
     index,
     record,
 ):
     # wait for the record to be indexed
     wait_for_index(
-        admin_client=session_admin_client,
+        admin_client=session_vector_client,
         namespace=DEFAULT_NAMESPACE,
         index=index
     )
