+++ tests/standard/test_vector_search.py
@@ -99,11 +99,10 @@ class vector_search_test_case:
 )
 def test_vector_search(
     session_vector_client,
-    session_admin_client,
     test_case,
 ):
     
-    session_admin_client.index_create(
+    session_vector_client.index_create(
         namespace=test_case.namespace,
         name=test_case.index_name,
         vector_field=test_case.vector_field,
@@ -132,7 +131,7 @@ def test_vector_search(
         )
     
     wait_for_index(
-        admin_client=session_admin_client,
+        admin_client=session_vector_client,
         namespace=test_case.namespace,
         index=test_case.index_name,
     )
@@ -154,7 +153,7 @@ def test_vector_search(
             key=key,
         )
 
-    session_admin_client.index_drop(
+    session_vector_client.index_drop(
         namespace=test_case.namespace,
         name=test_case.index_name,
     )
